+++ main.rb
@@ -1,7 +1,6 @@
 require './src/classes/app'
 
 def main
-  # puts File.exist?('./src/data/people.json')
   app = App.new
   app.console_entry_point
 end
