+++ src/data/people.json
@@ -1,11 +1,7 @@
 [
   {
-    "name": "dsfdf",
-    "age": 45
-  },
-  {
-    "name": "dfgs",
-    "age": 0,
-    "specialization": "dfgsd"
+    "name": "fhfh",
+    "age": 56,
+    "specialization": "fdg"
   }
 ]
